feat(api): 更新集群ID参数类型为Long，增强数据一致性
- 将多个控制器中的@ClusterId注解参数类型从Integer更改为Long，以确保集群ID的处理一致性。
- 更新相关方法的参数类型，确保与数据模型的兼容性，提升系统的稳定性与可维护性。

这些改动显著增强了API的健壮性与数据一致性，确保了整体设计的一致性。

+++ datasophon-api/src/main/java/com/datasophon/api/controller/v1/cluster/ClusterAlertGroupMapController.java
@@ -53,7 +53,7 @@ public class ClusterAlertGroupMapController {
      * 根据集群ID获取告警组映射列表
      */
     @RequestMapping("/list")
-    public Result<List<ClusterAlertGroupMapVO>> list(@ClusterId Integer clusterId) {
+    public Result<List<ClusterAlertGroupMapVO>> list(@ClusterId Long clusterId) {
         // 调用Service层方法，获取DTO列表
         List<ClusterAlertGroupMapDTO> dtoList = clusterAlertGroupMapService.getByClusterId(clusterId);
         // Controller层：DTO → VO转换
